Fishing Code Progress (Part2)
Fishing code can now work with fish prefab. More work needs to be done to implement raycast for controller instead of raycast from main cam. Fish needs to be moved to player after it is caught. Fish needs to be destroyed if player fails to catch it. setHook needs to be destroyed after minigame ends